requirements.txt
discord.py
dotenv
sqlalchemy
datetime
googletrans
langdetect
asyncio
psycopg[binary]